Finalize Uniforms and start on Vertex Attributes
This was meant to be two commits, but I forgot.

TODO: Vertex attributes need to track buffers to correctly throw errors
if buffers are modified in incompatible ways.

Uniform attributes are candidates for removal, as their management is
fairly complex and they do not fit with modern command-stream oriented
OpenGL (see what happens at glNext presentation).