Make buffer binding explicit
Explicitly bind buffers to specific targets, rather than setting a target to a
buffer and binding implicitly in methods.

Also removes the ability to bind one item several times, as few methods now bind
a buffer implicitly (and other items should be consistent). Perhaps add an
explicit context object with the ability to pop/push buffer bindings if this
becomes an issue.