Add support for parsing SPIRV
Roughing in an outline for parsing SPIRV programs. This needs to integrate with
the `datatypes` module, the OpCodes need to become a metaclass, so that you can
make instances of each OpCode, and the parsing module needs to be reorganized
into one for parsing the specification, and one for parsing programs.